送信部分を重複しないよう変更
