+++ frontend/app/(interview)/interview/[interviewId]/feedback/page.tsx
@@ -45,7 +45,7 @@ const FeedBack = () => {
 
     if (latestIds) {
       setFeedbackList(latestIds as UserAnswer[]);
-      console.log(feedbackList);
+      console.log(latestIds);
     }
   }
   return (
